+++ src/main/java/blackjack/domain/CardType.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain;
 
 public enum CardType {
     HEART("하트"),
