+++ src/test/java/blackjack/CardDeckTest.java
@@ -1,5 +1,7 @@
 package blackjack;
 
+import blackjack.domain.Card;
+import blackjack.domain.CardDeck;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
